Move all custom Exceptions to exceptions.py
