+++ pymc/step_methods/hmc/hmc.py
@@ -18,10 +18,11 @@ from typing import Any
 
 import numpy as np
 
+from pymc.exceptions import IntegrationError
 from pymc.stats.convergence import SamplerWarning
 from pymc.step_methods.compound import Competence
 from pymc.step_methods.hmc.base_hmc import BaseHMC, DivergenceInfo, HMCStepData
-from pymc.step_methods.hmc.integration import IntegrationError, State
+from pymc.step_methods.hmc.integration import State
 from pymc.vartypes import discrete_types
 
 __all__ = ["HamiltonianMC"]
