+++ tests/distributions/test_continuous.py
@@ -29,8 +29,8 @@ import pymc as pm
 
 from pymc.distributions.continuous import Normal, Uniform, get_tau_sigma, interpolated
 from pymc.distributions.dist_math import clipped_beta_rvs
+from pymc.exceptions import ParameterValueError
 from pymc.logprob.basic import icdf, logcdf, logp
-from pymc.logprob.utils import ParameterValueError
 from pymc.pytensorf import floatX
 from pymc.testing import (
     BaseTestDistributionRandom,
