+++ tests/distributions/test_truncated.py
@@ -24,11 +24,10 @@ from pymc.distributions.continuous import Exponential, Gamma, TruncatedNormalRV
 from pymc.distributions.shape_utils import change_dist_size
 from pymc.distributions.transforms import _default_transform
 from pymc.distributions.truncated import Truncated, TruncatedRV, _truncated
-from pymc.exceptions import TruncationError
+from pymc.exceptions import ParameterValueError, TruncationError
 from pymc.logprob.abstract import _icdf
 from pymc.logprob.basic import logcdf, logp
 from pymc.logprob.transforms import IntervalTransform
-from pymc.logprob.utils import ParameterValueError
 from pymc.testing import assert_moment_is_expected
